# swagger.yaml
openapi: 3.0.0
info:
  title: Invoice API
  description: API documentation for the Invoice Management System
  version: 1.0.0
  contact:
    email: support@example.com

servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://api.example.com/v1
    description: Production server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Invoice:
      type: object
      properties:
        _id:
          type: string
          format: uuid
        invoiceNumber:
          type: string
        customer:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceItem'
        totalAmount:
          type: number
        status:
          type: string
          enum: [draft, pending, partial, paid, overdue, cancelled]
        issueDate:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        notes:
          type: string
      required:
        - invoiceNumber
        - customer
        - items
        - dueDate

    InvoiceItem:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
          minimum: 0.01
        unitPrice:
          type: number
          minimum: 0
      required:
        - description
        - quantity
        - unitPrice

paths:
  /invoices:
    get:
      summary: Get all invoices
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, pending, partial, paid, overdue, cancelled]
      responses:
        '200':
          description: List of invoices
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  results:
                    type: integer
                  data:
                    type: object
                    properties:
                      invoices:
                        type: array
                        items:
                          $ref: '#/components/schemas/Invoice'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

    post:
      summary: Create new invoice
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Invoice'
      responses:
        '201':
          description: Invoice created
        '400':
          description: Invalid input
        '401':
          description: Unauthorized

  /invoices/{id}:
    get:
      summary: Get invoice by ID
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invoice found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '404':
          description: Invoice not found

    patch:
      summary: Update invoice
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Invoice'
      responses:
        '200':
          description: Invoice updated
        '400':
          description: Invalid input
        '404':
          description: Invoice not found

    delete:
      summary: Delete invoice
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Invoice deleted
        '404':
          description: Invoice not found 